升级.net8.0

+++ Backend/SCR/src/Services/Category/Category.Api/Models/CategoryType.cs
@@ -3,8 +3,8 @@ namespace Category.Api.Models;
 public class CategoryType
 {
     public string Name { get; set; }
-    public int Id { get; set; }
-    public int? ParentId { get; set; }
+    public long Id { get; set; }
+    public long? ParentId { get; set; }
     public DateTime CreatedDate { get; set; }
     public string Description { get; set; }
     public virtual CategoryType Parent { get; set; }
